requirements.txt
Scrapy==1.7.3
mysql-connector-python==8.0.18
configparser==4.0.2
coloredlogs==10.0
